requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
PyYAML>=6.0
pathlib2>=2.3.0

# Development and testing
pytest>=6.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Optional: For better performance
numba>=0.56.0

# Optional: For additional file formats
openpyxl>=3.0.0
xlrd>=2.0.0
pyarrow>=5.0.0

# Optional: For advanced visualization
bokeh>=2.4.0